requirements.txt
streamlit
openai
pandas
faiss-cpu
pypdf2